refactor(#39): rename state.rs to session_state.rs to avoid gsd::state collision
- git mv src/state.rs -> src/session_state.rs
- update mod decls in main.rs and lib.rs
- update refs in hook_handler.rs and utils.rs
- gsd::state left untouched (different domain)
